requirements.txt
openai~=1.0.0rc1
requests~=2.26.0
Pillow~=10.1.0
Flask~=2.2.0
ipstack~=0.1.4
bs4~=0.0.1
beautifulsoup4~=4.12.2
gnews~=0.3.5
googlesearch-python~=1.2.3
google-generativeai~=0.2.2
lxml~=4.9.3
```